refactored Loader component
