+++ apps/platform-app/src/app/[locale]/(moderation)/(routes)/moderation/loading.tsx
@@ -1,7 +1,12 @@
-import Loader from '@/components/Loader/Loader'
+import React from 'react'
+import { Loader } from '@gdh/ui-system'
+import { I18nNamespaces } from '@/i18n'
+import { useTranslations } from 'next-intl'
 
 const Loading = () => {
-  return <Loader />
+  const t = useTranslations(I18nNamespaces.Index)
+  return <Loader>{t('title')}</Loader>
 }
 
 export default Loading
+
