+++ apps/platform-app/src/app/loading.tsx
@@ -1,8 +1,11 @@
 import React from 'react'
-import Loader from '@/components/Loader/Loader'
+import { Loader } from '@gdh/ui-system'
+import { I18nNamespaces } from '@/i18n'
+import { useTranslations } from 'next-intl'
 
 const Loading = () => {
-  return <Loader />
+  const t = useTranslations(I18nNamespaces.Index)
+  return <Loader>{t('title')}</Loader>
 }
 
 export default Loading
