+++ packages/ui-system/src/components/index.ts
@@ -16,3 +16,4 @@ export { DropdownSearchInput } from "./Dropdowns/DropdownSearchInput/DropdownSea
 export { Logo } from "./Logo/Logo";
 export { TextArea } from "./TextArea/TextArea";
 export { ProfileViews } from "./my-profile/ProfileViews/ProfileViews"
+export { Loader } from "./Loader/Loader"
