+++ packages/ui-system/src/stories/Loader.stories.ts
@@ -1,9 +1,9 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import { ProfileViews } from "../index";
+import { Loader } from "../index";
 
 const meta = {
-  title: "Components/my-profile/ProfileViews",
-  component: ProfileViews,
+  title: "Components/Loader",
+  component: Loader,
   parameters: {
     // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/configure/story-layout
     layout: "centered",
@@ -12,11 +12,11 @@ const meta = {
   tags: ["autodocs"],
   // Use `fn` to spy on the onClick arg, which will appear in the actions panel once invoked: https://storybook.js.org/docs/essentials/actions#action-args
   args: {},
-} satisfies Meta<typeof ProfileViews>;
+} satisfies Meta<typeof Loader>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {
-  args: { viewCount: 14, children: "Profile Views" },
+  args: { children: "" },
 };
